Fixes #1581. Keys Pro7Sat1 fixup to transportid and networkid, rather than the traditonal networkid and serviceid match.
git-svn-id: http://svn.mythtv.org/svn/trunk@9690 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/eithelper.h
@@ -84,7 +84,7 @@ class EITHelper
 
     int                     gps_offset;
     uint                    sourceid;
-    QMap<uint,uint>         fixup;
+    QMap<uint64_t,uint>     fixup;
     ATSCSRCToEvents         incomplete_events;
     ATSCSRCToETTs           unmatched_etts;
